bug(server/client): get message but game no start

+++ api/src/main/java/akkamon/api/models/Position.java
@@ -1,10 +1,10 @@
 package akkamon.api.models;
 
 public class Position {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
