+++ domain/src/main/java/akkamon/domain/Akkamon.java
@@ -2,5 +2,5 @@ package akkamon.domain;
 
 public interface Akkamon {
     void newPlayerConnected(String name, String password);
-    void updateTrainerPosition(String name, int x, int y);
+    void updateTrainerPosition(String name, float x, float y);
 }
